Add View Model Factory and shared view model
