--- app/src/main/java/com/example/bakingapp/ui/recipes/RecipesFragmentViewModel.java
@@ -1,17 +1,0 @@
-package com.example.bakingapp.ui.recipes;
-
-import androidx.lifecycle.ViewModel;
-
-import com.example.bakingapp.data.Repository;
-import com.example.bakingapp.data.model.Recipe;
-
-import java.util.List;
-
-public class RecipesFragmentViewModel extends ViewModel {
-    private Repository repository = Repository.getInstance();
-    private List<Recipe> recipes = repository.getRecipes();
-
-    public List<Recipe> getRecipes() {
-        return recipes;
-    }
-}
