+++ app/src/main/java/com/example/bakingapp/ui/viewmodel/RecipeSharedViewModel.java
@@ -1,0 +1,35 @@
+package com.example.bakingapp.ui.viewmodel;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+import com.example.bakingapp.data.Repository;
+import com.example.bakingapp.data.model.Recipe;
+
+import java.util.List;
+
+public class RecipeSharedViewModel extends ViewModel {
+    private Repository repository;
+    private List<Recipe> recipes;
+    private MutableLiveData<Recipe> selectedRecipe = new MutableLiveData<>();
+
+    public RecipeSharedViewModel(Repository repository) {
+        this.repository = repository;
+    }
+
+    public List<Recipe> getRecipes() {
+        if (recipes == null) {
+            recipes = repository.getRecipes();
+        }
+        return recipes;
+    }
+
+    public LiveData<Recipe> getSelectedRecipe() {
+        return selectedRecipe;
+    }
+
+    public void setSelectedRecipe(Recipe recipe) {
+        selectedRecipe.setValue(recipe);
+    }
+}
