+++ app/src/main/res/layout-sw600dp/fragment_recipes.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.recipes.RecipesFragment">
+    tools:context=".ui.RecipesFragment">
 
     <GridView
         android:id="@+id/recipes_recycler_view"
