+++ app/src/main/res/layout/fragment_recipe_step.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.steps.RecipeStepFragment">
+    tools:context=".ui.RecipeStepFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recipe_steps_recycler_view"
