+++ app/src/main/res/layout/fragment_recipes.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="8dp"
-    tools:context=".ui.recipes.RecipesFragment">
+    tools:context=".ui.RecipesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recipes_recycler_view"
